requirements.txt
streamlit
google-generativeai
google-cloud-vision
pydantic
pdf2image
pillow
python-dotenv
fuzzywuzzy
python-Levenshtein
